# Add pipeline support to speed up large numbers of metrics