--- a/Functions/Send-DataDogMetric.ps1
+++ b/Functions/Send-DataDogMetric.ps1
@@ -28,19 +28,29 @@
     Send-DataDogMetric -Type Histogram -Name 'command.duration' -Value 12 -Tag @("command:my_command_name")
 .EXAMPLE
     Send-DataDogMetric -Type Gauge -Name 'random.value' -Value $randomvalue -ComputerName 192.168.0.1 -Port 8125
+.EXAMPLE
+    1..20000 | Send-DataDogMetric -Type Counter -Name 'incrementing.value' -Value { $_ }
+.EXAMPLE
+    Get-AppStatistics | Send-DataDogMetric -Type Gauge -Name 'appco.active_users' -Value { $_.ActiveUsers }
+.EXAMPLE
+    Get-Process |
+        Send-DataDogMetric -Name 'server.process.handles' -Type Counter -Value { $_.Handles } -Tag { @("process:$($_.ProcessName)","pid:$($_.Id)") }
 #>
 function Send-DataDogMetric {
     [CmdletBinding(SupportsShouldProcess=$true)]
     param(
-        [Parameter(Mandatory)]
+        [Parameter(Mandatory, ValueFromPipelineByPropertyName)]
+        [Alias('MetricName')]
         [ValidateNotNullOrEmpty()]
         [string]$Name,
 
-        [Parameter(Mandatory)]
+        [Parameter(Mandatory, ValueFromPipelineByPropertyName)]
+        [Alias('MetricValue')]
         [ValidateNotNullOrEmpty()]
         [string]$Value,
 
-        [Parameter(Mandatory)]
+        [Parameter(Mandatory, ValueFromPipelineByPropertyName)]
+        [Alias('MetricType')]
         [ValidateSet('Counter','Gauge','Histogram','Timer','Set','Distribution')]
         [string]$Type,
 
@@ -52,32 +62,50 @@ function Send-DataDogMetric {
         [ValidateRange(1,65535)]
         [int]$Port,
 
-        [Parameter()]
+        [Parameter(ValueFromPipelineByPropertyName)]
+        [Alias('MetricSampleRate')]
         [string]$SampleRate='1',
 
-        [Parameter()]
+        [Parameter(ValueFromPipelineByPropertyName)]
+        [Alias('MetricTag')]
+        [Alias('MetricTags')]
         [string[]]$Tag=@()
-
     )
-    $ddType = if ($Type -eq 'Timer') {
-        'ms'
-    }
-    else {
-        $Type.ToLower()[0]
+
+    Begin {
+        $statsdParams = @{
+            Verbose = $VerbosePreference -as [bool]
+            Debug = $DebugPreference -as [bool]
+        }
+        if ($ComputerName) {
+            $statsdParams.ComputerName = $ComputerName
+        }
+        if ($Port) {
+            $statsdParams.Port = $Port
+        }
+
+        # start a steppable pipeline for Send-StatsD so we can ensure we use a single UDP socket across the pipeline
+        $sendStatsD = { Send-StatsD @statsdParams }.GetSteppablePipeline($MyInvocation.CommandOrigin)
+        $sendStatsD.Begin($true)
     }
 
-    $data = "$($Name):$($Value)|$ddType|@$SampleRate|#$([string]::Join(',',$Tag))"
-    $statsdParams = @{
-        Data = $data
-    }
-    if ($ComputerName) {
-        $statsdParams.ComputerName = $ComputerName
-    }
-    if ($Port) {
-        $statsdParams.Port = $Port
+    Process {
+        $ddType = if ($Type -eq 'Timer') {
+            'ms'
+        }
+        else {
+            $Type.ToLower()[0]
+        }
+
+        $tagData = [string]::Join(',', $Tag)
+        $data = "${Name}:${Value}|${ddType}|@${SampleRate}|#${tagData}"
+
+        if ($PSCmdlet.ShouldProcess($data)) {
+            $sendStatsD.Process($data)
+        }
     }
 
-    if ($PSCmdlet.ShouldProcess("Sending DataDog metric: $data")) {
-        Send-StatsD @statsdParams
+    End {
+        $sendStatsD.End()
     }
 }