## Changes
* Update the existing pipeline support in `Send-StatsD` to use a single `UDPClient` for the entire pipeline, instead creating and destroying one for each pipeline item
* Update `Send-DataDogMetric` to accept its parameters via pipeline object (by property name). This unlocks 2 things:
  * Using pipeline input to set all of the relevant parameters (via PowerShell's ability to use a scriptblock for any pipeline-bound param)
  * Use the new single `UDPClient` support in `Send-StatsD` to get the same performance benefits when sending lots of metrics.
* Add new pipeline examples to `Send-DataDogMetric` and include them in the README

## Speed Improvement
By using the pipeline this way, time taken was reduced significantly. Examples with 20 000 iterations:

**Not using pipeline: 12.946 seconds**
```powershell
Measure-Command -Expression { (1..20000).ForEach({ Send-DataDogMetric -Name 'a' -Type Counter -Tag @("ABC:$_") -Value $_ }) }
```

**Using pipeline: 3.898 seconds**
```powershell
Measure-Command -Expression { 1..20000 | Send-DataDogMetric -Name 'a' -Type Counter -Tag { @("ABC:$_") } -Value { $_ } }
```

For `Send-StatsD` directly:

**Not using pipeline: 9.104s**
```powershell
Measure-Command -Expression { (1..20000).ForEach({ Send-StatsD -Data $_ }) }
```

**Using pipeline: 2.642s**
```powershell
Measure-Command -Expression { 1..20000 | Send-StatsD }
```
---

I didn't update `Send-DataDogEvent` with pipeline support as it seemed like something that isn't likely to be called as often, but I don't use it myself (yet), so if that's desired I could add it in in another PR.